third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/caret-color-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'caret-color' to the 'currentcolor' keyword: currentcolor
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'caret-color' to the 'auto' keyword: auto
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
Harness: the test ran to completion.
